mesh-studio/task-manager-api/requirements.txt
pyjwt
boto3
bcrypt
passlib
pylint 
black 
# moto
# pytest